improve html reporter and add stdout and json reporters
