--- src/reporters/css.css
@@ -1,210 +1,0 @@
-:root {
-    --color-bg: white;
-    --color-grey: #ddd;
-    --color-bg-muted: rgb(246, 248, 250);
-    --color-file-info-bg: #f1f8ff;
-    --color-font-muted: #888;
-    --color-error: indianred;
-    --color-warn: #ccb021;
-    --width-container: 1200px;
-    --width-gutter: 20px;
-}
-
-body > header {
-    background: cornflowerblue;
-    color: white;
-    padding: 1rem;
-    margin-bottom: 1rem;
-    display: flex;
-    justify-content: space-between;
-}
-
-body {
-    font-family: Verdana, Arial, sans-serif;
-    line-height: 1.6;
-    margin: 0;
-    padding: 0;
-    background-color: var(--color-bg);
-}
-
-.sidebar {
-    margin-right: 1rem;
-    max-width: 18rem;
-}
-
-.sidebar header {
-    background: var(--color-file-info-bg);
-    padding: 0.5rem;
-    border-bottom: 1px solid var(--color-grey);
-}
-
-.sidebar header > p {
-    margin: 0;
-}
-
-.sidebar-content {
-    padding: 0.5rem;
-}
-
-.stat {
-    display: flex;
-    justify-content: space-between;
-    font-size: 14px;
-}
-
-.stat p {
-    margin: 0;
-}
-
-.sidebar > div {
-    margin-bottom: 1rem;
-}
-
-.sidebar input#search {
-    border: 1px solid var(--color-grey);
-    padding: 0.5rem;
-    width: 12rem;
-}
-
-.sidebar-content > div {
-    margin-bottom: 1rem;
-}
-
-.container {
-    margin: 0 auto;
-    width: var(--width-container);
-    display: flex;
-}
-
-.file-info {
-    padding: 0.5rem;
-    background: var(--color-file-info-bg);
-    display: flex;
-    justify-content: space-between;
-    cursor: pointer;
-    user-select: none;
-}
-
-.file-info svg {
-    color: var(--color-font-muted);
-}
-
-.file-info p {
-    margin: 0;
-}
-
-.file-info, .error, table, .warn, .sidebar > div {
-    border: 1px solid var(--color-grey);
-}
-
-.file-info.error, .error {
-    border-left: 5px solid var(--color-error);
-}
-.file-info.warn, .warn {
-    border-left: 5px solid var(--color-warn);
-}
-
-.file {
-    margin-bottom: 1rem;
-}
-
-table {
-    border-collapse: collapse;
-    border-top: 0;
-}
-
-.gutter {
-    width: var(--width-gutter);
-    line-height: 1.4;
-    font-family: monospace;
-    border-right: 1px solid var(--color-grey);
-    text-align: right;
-    padding: 0 0.5rem;
-    background: var(--color-bg-muted);
-    vertical-align: top;
-}
-
-.line {
-    font-family: monospace;
-    padding: 0 0.5rem;
-    width: calc(var(--width-container) - var(--width-gutter));
-}
-
-.line p {
-    margin: 0;
-}
-
-td.has-error > p {
-    text-decoration: var(--color-error) wavy underline;
-}
-
-.keyword {
-    color: cornflowerblue;
-}
-
-td .error, td .warn {
-    vertical-align: middle;
-    padding: 0.5rem;
-    margin: 0.5rem 0;
-}
-
-td .error svg {
-    color: var(--color-error);
-    width: 2rem;
-}
-td .warn svg {
-    color: var(--color-warn);
-    width: 2rem;
-}
-
-td .error > header p, td .warn > header p {
-    display: inline-block;
-    vertical-align: top;
-    margin: 0.1rem 0 0;
-}
-
-.rule-name, .text-muted {
-    color: var(--color-font-muted);
-}
-
-td .error .line-info {
-    color: var(--color-error);
-    font-weight: 600;
-}
-td .warn .line-info {
-    color: var(--color-warn);
-    font-weight: 600;
-}
-
-.error-count, .warn-count {
-    border-radius: 100%;
-    color: white;
-    padding: 0 0.5rem;
-}
-.error-count {
-    background: var(--color-error);
-}
-.warn-count {
-    background: var(--color-warn);
-}
-
-.file .file-content {
-    display: none;
-}
-
-.file.active .file-content {
-    display: block;
-}
-
-.up-arrow {
-    display: none;
-}
-.file.active .up-arrow {
-    display: block;
-}
-.file.active .down-arrow {
-    display: none;
-}
-.highlight {
-    background: yellow;
-}
