+++ tests/acceptance/features/no-unnamed-scenarios/invalid.feature
@@ -1,6 +1,9 @@
 Feature: No Unnamed Scenarios
 
+  This is a description block
+
   Scenario:
+    # This is a comment
     Given I do something
     When I do something else
     Then I should have done something
